# gateway/kitty-chat/src/lib/gen/openapi.json
{
  "components": {
    "schemas": {
...
      "BuildStartRequest": {
        "properties": {
          "auto_approve": {
            "default": false,
            "title": "Auto Approve",
            "type": "boolean"
          },
          "goal": {
            "maxLength": 3000,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "target_dir": {
            "default": "",
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "goal"
        ],
        "title": "BuildStartRequest",
        "type": "object"
      },
...
      "DeployRequest": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "platform": {
            "default": "docker",
            "title": "Platform",
            "type": "string"
          },
          "target_dir": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Target Dir",
            "type": "string"
          }
        },
        "required": [
          "target_dir"
        ],
        "title": "DeployRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationsBuildBuildIdApproveStageResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsBuildBuildIdApproveStageResponse",
        "type": "object"
      },
      "IntegrationsBuildBuildIdResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsBuildBuildIdResponse",
        "type": "object"
      },
      "IntegrationsBuildStartResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsBuildStartResponse",
        "type": "object"
      },
      "IntegrationsBuildsResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsBuildsResponse",
        "type": "object"
      },
      "IntegrationsDeployResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsDeployResponse",
        "type": "object"
      },
...
      "IntegrationsImessageRecentResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsImessageRecentResponse",
        "type": "object"
      },
      "IntegrationsImessageSendResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsImessageSendResponse",
        "type": "object"
      },
...
      "IntegrationsNudgeNudgeIdDismissResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsNudgeNudgeIdDismissResponse",
        "type": "object"
      },
...
      "IntegrationsPluginNameDisableResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsPluginNameDisableResponse",
        "type": "object"
      },
      "IntegrationsPluginNameEnableResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsPluginNameEnableResponse",
        "type": "object"
      },
...
      "IntegrationsVerifyResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "IntegrationsVerifyResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyRequest": {
        "properties": {
          "target_dir": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Target Dir",
            "type": "string"
          },
          "test_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Path"
          }
        },
        "required": [
          "target_dir"
        ],
        "title": "VerifyRequest",
        "type": "object"
      },
...
      "iMessageSendRequest": {
        "properties": {
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "recipient": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Recipient",
            "type": "string"
          }
        },
        "required": [
          "recipient",
          "message"
        ],
        "title": "iMessageSendRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kitty Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/build/start": {
      "post": {
        "operationId": "build_start_build_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsBuildStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Start",
        "tags": [
          "integrations"
        ]
      }
    },
    "/build/{build_id}": {
      "get": {
        "operationId": "build_status_build__build_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsBuildBuildIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Status",
        "tags": [
          "integrations"
        ]
      }
    },
    "/build/{build_id}/approve/{stage}": {
      "post": {
        "operationId": "build_approve_build__build_id__approve__stage__post",
        "parameters": [
          {
            "in": "path",
            "name": "build_id",
            "required": true,
            "schema": {
              "title": "Build Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stage",
            "required": true,
            "schema": {
              "title": "Stage",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsBuildBuildIdApproveStageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Approve",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/builds": {
      "get": {
        "operationId": "build_list_builds_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsBuildsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build List",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/deploy": {
      "post": {
        "operationId": "deploy_project_deploy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsDeployResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deploy Project",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/imessage/recent": {
      "get": {
        "operationId": "imessage_recent_imessage_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsImessageRecentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Imessage Recent",
        "tags": [
          "integrations"
        ]
      }
    },
    "/imessage/send": {
      "post": {
        "operationId": "imessage_send_imessage_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/iMessageSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsImessageSendResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Imessage Send",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/nudge/{nudge_id}/dismiss": {
      "post": {
        "operationId": "nudge_dismiss_nudge__nudge_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "nudge_id",
            "required": true,
            "schema": {
              "title": "Nudge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsNudgeNudgeIdDismissResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nudge Dismiss",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/plugin/{name}/disable": {
      "post": {
        "operationId": "plugin_disable_plugin__name__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsPluginNameDisableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plugin Disable",
        "tags": [
          "integrations"
        ]
      }
    },
    "/plugin/{name}/enable": {
      "post": {
        "operationId": "plugin_enable_plugin__name__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsPluginNameEnableResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plugin Enable",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/verify": {
      "post": {
        "operationId": "verify_run_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsVerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Run",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}